feature: Filterable instances // fix circular ref
